Make Roster and Matches interfaces consistent
This commit adjusts both the Matches and Roster interface and storage to
work consistently.  Both are now stored in the ETS table as a tuple with
all fields as defined in the event CSV.  The set/1 method now expects a
tuple and the list method now returns the same tuples.